feat(esl-share): retire esl-share and rename esl-share-popup-trigger into esl-share
