+++ pages/src/esl-share/esl-share.ts
@@ -1,8 +1,9 @@
 import '@exadel/esl/modules/esl-share/buttons/all';
-import {ESLShare, ESLShareConfig} from '@exadel/esl/modules/esl-share/core';
+import {ESLShare, ESLShareList, ESLShareConfig} from '@exadel/esl/modules/esl-share/core';
 
 ESLShareConfig.append({name: 'site', 'list': 'facebook twitter linkedin telegram copy'});
 ESLShare.register();
+ESLShareList.register();
 
 ESLShareConfig.set(
   fetch('/assets/share/config.json').then((response) => response.json())
