+++ src/modules/esl-share/core.less
@@ -1,3 +1,3 @@
 @import './core/esl-share-button.less';
 @import './core/esl-share-list.less';
-@import './core/esl-share-popup-trigger.less';
+@import './core/esl-share.less';
