+++ src/modules/esl-share/core.ts
@@ -5,4 +5,3 @@ export * from './core/esl-share-button';
 export * from './core/esl-share-config';
 export * from './core/esl-share-list';
 export * from './core/esl-share-popup';
-export * from './core/esl-share-popup-trigger';
