+++ src/modules/esl-share/core/esl-share.less
@@ -1,4 +1,4 @@
-esl-share-popup-trigger {
+esl-share {
   display: inline-block;
   cursor: pointer;
 }
