+++ src/modules/esl-share/core/esl-share.shape.ts
@@ -1,22 +1,22 @@
-import type {ESLBaseElementShape} from '../../esl-base-element/core/esl-base-element.shape';
+import type {ESLTriggerTagShape} from '../../esl-trigger/core/esl-trigger.shape';
 import type {ESLShare} from './esl-share';
 
 /**
- * Tag declaration interface of ESL Share element
+ * Tag declaration interface of ESLSharePopupTrigger element
  * Used for TSX declaration
  */
-export interface ESLShareTagShape extends ESLBaseElementShape<ESLShare> {
-  /** Define the list of social networks or groups of them to display (all by default) */
+export interface ESLShareTagShape<T extends ESLShare = ESLShare> extends ESLTriggerTagShape<T> {
+  /** Define the list of social networks or groups of them to display at popup (all by default) */
   list?: string;
-  /** Define share buttons rendering mode */
-  mode?: string;
   /** Define URL to share (current page URL by default) */
   'share-url'?: string;
   /** Define title to share (current document title by default) */
   'share-title'?: string;
+  /** Define JSON of initial params to pass into the popup */
+  'popup-initial-params'?: string;
 
-  /** Children are not allowed for ESLShare */
-  children?: never[];
+  /** Allowed children */
+  children?: any;
 }
 
 declare global {
